requirements.txt
# Core ML dependencies
torch
torchvision
transformers>=4.50.0

# Qwen VL dependencies
qwen-vl-utils
accelerate

# Image processing
Pillow
scikit-image
scikit-learn

# API clients (optional, for cloud models)
openai
google-genai
google-generativeai
google-api-core

# Utilities
python-liquid
pyyaml
numpy
matplotlib